backend/requirements.txt
Flask
mysql-connector-python
python-dotenv
Flask-RESTful
Flask-JWT-Extended